refactor(layout): simplify layout component
